+++ HISTORY.md
@@ -1,6 +1,25 @@
 # History
 
 
+# 2024-02-01, 12.3.1
+
+- Improved the typings of the arguments of `ArrayNode`, `FunctionNode`, 
+  `IndexNode`, `OperatorNode`, and `RelationalNode` (#3123). Thanks @sylee957.
+- Added a fully featured code editor example with CodeMirror and Katex (#3027).
+  Thanks @dvd101x.
+- Fix: #3114 build warnings related to a number of wrong `/* #__PURE__ */` 
+  annotations.
+- Fix: #3142 support BigNumber values for the options of function `format`:
+  `precision`, `wordSize`, `lowerExp`, `upperExp`. Support BigNumber values
+  for the option `wordSize` in the functions `hex`, `bin`, and `oct`. 
+- Fix: #3125 type definitions of function `hypot` (#3144). 
+  Thanks @silentmissile.
+- Fix: #3141 `help(config)` altering the actual `config` when evaluating the 
+  examples.
+- Docs: #3145 fix documentation about REPL, it does require a build step 
+  nowadays.
+
+
 # 2024-01-12, 12.3.0
 
 - Implement support new metric prefixes: `ronna` (`R`), `quetta` (`Q`),